:recycle: Move utility functions to repl_helper Use static ops
