+++ src/repl.rs
@@ -1,156 +1,25 @@
-use dynasmrt::{dynasm, x64::Assembler, DynamicLabel, DynasmApi, DynasmLabelApi};
+use dynasmrt::{dynasm, x64::Assembler, DynamicLabel, DynasmApi};
 use im::hashmap;
 use im::HashMap;
-use im::HashSet;
+use once_cell::sync::Lazy;
 use sexp::*;
 use std::io::Write;
 use std::mem;
 use std::panic;
+use std::sync::Mutex;
 
 use crate::compiler::*;
 use crate::parser::*;
+use crate::repl_helper::*;
 use crate::structs::*;
 
-fn snek_error_exit(errcode: i64) {
-    // print error message according to writeup
-    eprintln!(
-        "an error ocurred {errcode}: {}",
-        match errcode {
-            1 => "invalid representation",
-            i if i >= 20 && i <= 29 => "invalid argument",
-            i if i >= 30 && i <= 39 => "overflow",
-            i if i == 40 => "index out of range",
-            _ => "",
-        }
-    );
-    std::process::exit(1);
-}
+static OPS: Lazy<Mutex<Assembler>> = Lazy::new(|| Mutex::new(Assembler::new().unwrap()));
 
-fn snek_error_print(errcode: i64) {
-    // print error message according to writeup
-    eprintln!(
-        "an error ocurred {errcode}: {}",
-        match errcode {
-            1 => "invalid representation",
-            i if i >= 20 && i <= 29 => "invalid argument",
-            i if i >= 30 && i <= 39 => "overflow",
-            i if i == 40 => "index out of range",
-            _ => "",
-        }
-    );
-}
-
-fn add_interface_calls(ops: &mut Assembler, lbls: &mut HashMap<Label, DynamicLabel>, exit: bool) {
-    let snek_error_lbl = ops.new_dynamic_label();
-    lbls.insert(Label::new(Some("snek_error_stub")), snek_error_lbl);
-    dynasm!(ops; .arch x64; =>snek_error_lbl);
-    if exit {
-        dynasm!(ops; .arch x64; mov rax, QWORD snek_error_exit as _);
-    } else {
-        dynasm!(ops; .arch x64; mov rax, QWORD snek_error_print as _);
-    }
-    dynasm!(ops;
-        .arch x64;
-        mov rsp, [r15 + 8];
-        call rax;
-        ret
-    );
-
-    let snek_print_lbl = ops.new_dynamic_label();
-    lbls.insert(Label::new(Some("snek_print")), snek_print_lbl);
-    dynasm!(ops; .arch x64; =>snek_print_lbl; mov rax, QWORD print_result as _; call rax; ret);
-
-    let snek_deep_equal_lbl = ops.new_dynamic_label();
-    lbls.insert(Label::new(Some("snek_deep_equal")), snek_deep_equal_lbl);
-    dynasm!(ops; .arch x64; =>snek_deep_equal_lbl; mov rax, QWORD deep_equal as _; call rax; ret);
-}
-
-fn parse_input(input: &str) -> (i64, Option<Type>) {
-    // parse the input string into internal value representation
-    match input {
-        "true" => (TRUE_VAL, Some(Type::Bool)),
-        "false" | "" => (FALSE_VAL, Some(Type::Bool)),
-        _ => (
-            (input.parse::<i64>().expect("Invalid") as i64) << 1,
-            Some(Type::Int),
-        ),
-    }
-}
-
-fn deep_equal_recurse(l: i64, r: i64, seen: &mut HashSet<(i64, i64)>) -> bool {
-    // If not list, early exit
-    if l & 3 != 1 || r & 3 != 1 || l == NIL_VAL || r == NIL_VAL {
-        return l == r;
-    }
-
-    if seen.contains(&(l, r)) {
-        return true;
-    }
-    seen.insert((l, r));
-
-    let la = (l - 1) as *const i64;
-    let ra = (r - 1) as *const i64;
-    let lc = unsafe { *la } as isize;
-    let rc = unsafe { *ra } as isize;
-    // Check length
-    if lc != rc {
-        return false;
-    }
-    for i in 1..=lc {
-        let ln = unsafe { *la.offset(i) };
-        let rn = unsafe { *ra.offset(i) };
-        if !deep_equal_recurse(ln, rn, seen) {
-            return false;
-        }
-    }
-    return true;
-}
-
-fn deep_equal(l: i64, r: i64) -> i64 {
-    if deep_equal_recurse(l, r, &mut HashSet::new()) {
-        TRUE_VAL
-    } else {
-        FALSE_VAL
-    }
-}
-
-fn snek_str(val: i64, seen: &mut HashSet<i64>) -> String {
-    if val == TRUE_VAL {
-        "true".to_string()
-    } else if val == FALSE_VAL {
-        "false".to_string()
-    } else if val % 2 == 0 {
-        format!("{}", val >> 1)
-    } else if val == NIL_VAL {
-        "nil".to_string()
-    } else if val & 3 == 1 {
-        if seen.contains(&val) {
-            return "(list <cyclic>)".to_string();
-        }
-        seen.insert(val);
-        let addr = (val - 1) as *const i64;
-        let count = unsafe { *addr } as usize;
-        let mut v: Vec<i64> = vec![0; count];
-        for i in 1..=count {
-            v[i - 1] = unsafe { *addr.offset(i as isize) };
-        }
-        let result = format!(
-            "(list {})",
-            v.iter()
-                .map(|x| snek_str(*x, seen))
-                .collect::<Vec<String>>()
-                .join(" ")
-        );
-        seen.remove(&val);
-        return result;
-    } else {
-        format!("Unknown value: {}", val)
-    }
-}
-
-fn print_result(result: i64) -> i64 {
-    println!("{}", snek_str(result, &mut HashSet::new()));
-    return result;
+#[macro_export]
+macro_rules! stat {
+    ($var: ident) => {
+        &mut $var.lock().unwrap()
+    };
 }
 
 fn eval(
@@ -162,9 +31,7 @@ fn eval(
     instrs_to_asm(&instrs, ops, labels);
     dynasm!(ops; .arch x64; ret);
 
-    if let Err(e) = ops.commit() {
-        panic!("error committing ops {e}");
-    }
+    ops.commit().unwrap();
 
     let jitted_fn: extern "C" fn() -> i64 = {
         let reader = ops.reader();
@@ -188,14 +55,13 @@ pub fn repl(eval_input: Option<(&Vec<Expr>, &Expr, &str)>) {
     let mut com = ContextMut::new();
     com.curr_heap_ptr = heap.as_mut_ptr() as i64;
 
-    let mut ops = Assembler::new().unwrap();
     let mut labels: HashMap<Label, DynamicLabel> = hashmap! {};
 
     let mut input = (FALSE_VAL, Some(Type::Bool));
 
     // Eval
     if let Some((eval_fns, eval_in, input_str)) = eval_input {
-        add_interface_calls(&mut ops, &mut labels, true);
+        add_interface_calls(stat!(OPS), &mut labels, true);
         let mut instrs: Vec<Instr> = vec![Instr::Mov(MovArgs::ToReg(
             Reg::R15,
             Arg64::Imm64(heap.as_mut_ptr() as i64),
@@ -220,18 +86,16 @@ pub fn repl(eval_input: Option<(&Vec<Expr>, &Expr, &str)>) {
         labels.extend(com.fns.iter().fold(hashmap! {}, |mut acc, (f, _)| {
             acc.insert(
                 Label::new(Some(&format!("fun_{f}"))),
-                ops.new_dynamic_label(),
+                stat!(OPS).new_dynamic_label(),
             );
             acc
         }));
         instrs_to_asm(
             &compile_func_defns(eval_fns, &mut com),
-            &mut ops,
+            stat!(OPS),
             &mut labels,
         );
-        if let Err(e) = ops.commit() {
-            println!("{e}");
-        }
+        stat!(OPS).commit().unwrap();
 
         // Setup input
         input = parse_input(input_str);
@@ -243,12 +107,12 @@ pub fn repl(eval_input: Option<(&Vec<Expr>, &Expr, &str)>) {
             Some(&mut com),
             input.1,
         ));
-        print_result(eval(&mut ops, &mut labels, &mut instrs));
+        print_result(eval(stat!(OPS), &mut labels, &mut instrs));
         return;
     }
 
     // REPL
-    add_interface_calls(&mut ops, &mut labels, false);
+    add_interface_calls(stat!(OPS), &mut labels, false);
     let mut line = String::new();
     loop {
         println!("{co:?}");
@@ -376,10 +240,8 @@ pub fn repl(eval_input: Option<(&Vec<Expr>, &Expr, &str)>) {
                         },
                     );
 
-                    instrs_to_asm(&instrs, &mut ops, &mut labels);
-                    if let Err(e) = ops.commit() {
-                        println!("{e}");
-                    }
+                    instrs_to_asm(&instrs, stat!(OPS), &mut labels);
+                    stat!(OPS).commit().unwrap();
                     // Do not run any code
                     for i in &*instrs {
                         println!("{i:?}");
@@ -403,7 +265,7 @@ pub fn repl(eval_input: Option<(&Vec<Expr>, &Expr, &str)>) {
             for i in &*instrs {
                 println!("{i:?}");
             }
-            print_result(eval(&mut ops, &mut labels, instrs));
+            print_result(eval(stat!(OPS), &mut labels, instrs));
         };
 
         // Increase define stack if needed
