+++ src/repl_helper.rs
@@ -1,0 +1,152 @@
+use dynasmrt::{dynasm, x64::Assembler, DynamicLabel, DynasmApi, DynasmLabelApi};
+use im::HashMap;
+
+use crate::compiler::*;
+use crate::structs::*;
+use im::HashSet;
+
+pub fn snek_error_exit(errcode: i64) {
+    // print error message according to writeup
+    eprintln!(
+        "an error ocurred {errcode}: {}",
+        match errcode {
+            1 => "invalid representation",
+            i if i >= 20 && i <= 29 => "invalid argument",
+            i if i >= 30 && i <= 39 => "overflow",
+            i if i == 40 => "index out of range",
+            _ => "",
+        }
+    );
+    std::process::exit(1);
+}
+
+pub fn snek_error_print(errcode: i64) {
+    // print error message according to writeup
+    eprintln!(
+        "an error ocurred {errcode}: {}",
+        match errcode {
+            1 => "invalid representation",
+            i if i >= 20 && i <= 29 => "invalid argument",
+            i if i >= 30 && i <= 39 => "overflow",
+            i if i == 40 => "index out of range",
+            _ => "",
+        }
+    );
+}
+
+pub fn add_interface_calls(
+    ops: &mut Assembler,
+    lbls: &mut HashMap<Label, DynamicLabel>,
+    exit: bool,
+) {
+    let snek_error_lbl = ops.new_dynamic_label();
+    lbls.insert(Label::new(Some("snek_error_stub")), snek_error_lbl);
+    dynasm!(ops; .arch x64; =>snek_error_lbl);
+    if exit {
+        dynasm!(ops; .arch x64; mov rax, QWORD snek_error_exit as _);
+    } else {
+        dynasm!(ops; .arch x64; mov rax, QWORD snek_error_print as _);
+    }
+    dynasm!(ops;
+        .arch x64;
+        mov rsp, [r15 + 8];
+        call rax;
+        ret
+    );
+
+    let snek_print_lbl = ops.new_dynamic_label();
+    lbls.insert(Label::new(Some("snek_print")), snek_print_lbl);
+    dynasm!(ops; .arch x64; =>snek_print_lbl; mov rax, QWORD print_result as _; call rax; ret);
+
+    let snek_deep_equal_lbl = ops.new_dynamic_label();
+    lbls.insert(Label::new(Some("snek_deep_equal")), snek_deep_equal_lbl);
+    dynasm!(ops; .arch x64; =>snek_deep_equal_lbl; mov rax, QWORD deep_equal as _; call rax; ret);
+}
+
+pub fn parse_input(input: &str) -> (i64, Option<Type>) {
+    // parse the input string into internal value representation
+    match input {
+        "true" => (TRUE_VAL, Some(Type::Bool)),
+        "false" | "" => (FALSE_VAL, Some(Type::Bool)),
+        _ => (
+            (input.parse::<i64>().expect("Invalid") as i64) << 1,
+            Some(Type::Int),
+        ),
+    }
+}
+
+pub fn deep_equal_recurse(l: i64, r: i64, seen: &mut HashSet<(i64, i64)>) -> bool {
+    // If not list, early exit
+    if l & 3 != 1 || r & 3 != 1 || l == NIL_VAL || r == NIL_VAL {
+        return l == r;
+    }
+
+    if seen.contains(&(l, r)) {
+        return true;
+    }
+    seen.insert((l, r));
+
+    let la = (l - 1) as *const i64;
+    let ra = (r - 1) as *const i64;
+    let lc = unsafe { *la } as isize;
+    let rc = unsafe { *ra } as isize;
+    // Check length
+    if lc != rc {
+        return false;
+    }
+    for i in 1..=lc {
+        let ln = unsafe { *la.offset(i) };
+        let rn = unsafe { *ra.offset(i) };
+        if !deep_equal_recurse(ln, rn, seen) {
+            return false;
+        }
+    }
+    return true;
+}
+
+pub fn deep_equal(l: i64, r: i64) -> i64 {
+    if deep_equal_recurse(l, r, &mut HashSet::new()) {
+        TRUE_VAL
+    } else {
+        FALSE_VAL
+    }
+}
+
+pub fn snek_str(val: i64, seen: &mut HashSet<i64>) -> String {
+    if val == TRUE_VAL {
+        "true".to_string()
+    } else if val == FALSE_VAL {
+        "false".to_string()
+    } else if val % 2 == 0 {
+        format!("{}", val >> 1)
+    } else if val == NIL_VAL {
+        "nil".to_string()
+    } else if val & 3 == 1 {
+        if seen.contains(&val) {
+            return "(list <cyclic>)".to_string();
+        }
+        seen.insert(val);
+        let addr = (val - 1) as *const i64;
+        let count = unsafe { *addr } as usize;
+        let mut v: Vec<i64> = vec![0; count];
+        for i in 1..=count {
+            v[i - 1] = unsafe { *addr.offset(i as isize) };
+        }
+        let result = format!(
+            "(list {})",
+            v.iter()
+                .map(|x| snek_str(*x, seen))
+                .collect::<Vec<String>>()
+                .join(" ")
+        );
+        seen.remove(&val);
+        return result;
+    } else {
+        format!("Unknown value: {}", val)
+    }
+}
+
+pub fn print_result(result: i64) -> i64 {
+    println!("{}", snek_str(result, &mut HashSet::new()));
+    return result;
+}
